+++ src/main/java/frc/robot/commands/Autonomous.java
@@ -18,7 +18,7 @@ public class Autonomous extends CommandBase {
 
     @Override
     public void initialize() { 
-        VisionProcessing.get_vision_vectors();
+
     }
 
     @Override   
@@ -37,20 +37,9 @@ public class Autonomous extends CommandBase {
     }
 
     public void fake_auto() { 
-      SmartDashboard.putString("Bazinga", "Punk");
-
-      //THIS SECTION OF CODE CAN BE MOVED!!
-      //Tells the robot to face the targeted ball
-      //Note: At this time, I do not fully understand how to make 
-      //motors move, and do not want to screw anything up, so I will 
-      //leave placeholders for those actions
-      VisionProcessing.get_vision_vectors();
-      VisionProcessing.get_radius();
-      SmartDashboard.putNumber("Center X", VisionProcessing.get_center("X"));
-      SmartDashboard.putNumber("Center Y", VisionProcessing.get_center("Y"));
       //Code here to make motors spin the robot counterclockwise
 
-    /*
+    
         int i = 0; 
         while (i < 2000) { 
             drive_train.tankDrive(-0.7, 0.7);
@@ -60,6 +49,6 @@ public class Autonomous extends CommandBase {
 
         System.out.println("ENDED!"); 
         isFinished();
-    */
+    
     }
 }
